Make control comments editable

+++ apps/frontend/src/components/global/sidebaritems/SidebarFileList.vue
@@ -13,7 +13,18 @@
     </v-list-item-avatar>
 
     <v-list-item-content>
-      <v-list-item-title>{{ file.filename }}</v-list-item-title>
+      <v-list-item-title>
+        {{ file.filename }}
+        <v-icon
+          v-if="file.hasUnsavedChanges"
+          class="ml-2"
+          color="warning"
+          small
+          title="Unsaved comments edits"
+        >
+          mdi-content-save-alert
+        </v-icon>
+      </v-list-item-title>
     </v-list-item-content>
 
     <v-list-item-action v-if="serverMode" @click.stop="save_file">
@@ -36,7 +47,7 @@
 import RouteMixin from '@/mixins/RouteMixin';
 import ServerMixin from '@/mixins/ServerMixin';
 import {FilteredDataModule} from '@/store/data_filters';
-import {InspecDataModule} from '@/store/data_store';
+import {InspecDataModule, UNSAVED_CHANGES_MESSAGE} from '@/store/data_store';
 import {EvaluationModule} from '@/store/evaluations';
 import {EvaluationFile, ProfileFile} from '@/store/report_intake';
 import {SnackbarModule} from '@/store/snackbar';
@@ -75,6 +86,13 @@ export default class SidebarFileList extends mixins(ServerMixin, RouteMixin) {
 
   //removes uploaded file from the currently observed files
   remove_file() {
+    if (
+      this.file.hasUnsavedChanges &&
+      !globalThis.confirm(`${UNSAVED_CHANGES_MESSAGE}\n\nRemove it anyway?`)
+    ) {
+      return;
+    }
+
     EvaluationModule.removeEvaluation(this.file.uniqueId);
     InspecDataModule.removeFile(this.file.uniqueId);
     // Remove any database files that may have been in the URL
@@ -86,7 +104,7 @@ export default class SidebarFileList extends mixins(ServerMixin, RouteMixin) {
   //saves file to database
   save_file() {
     if (this.file?.database_id) {
-      SnackbarModule.failure('This file is already in the database.');
+      this.update_database_file(this.file);
     } else if (this.file) {
       this.save_to_database(this.file);
     }
@@ -94,7 +112,43 @@ export default class SidebarFileList extends mixins(ServerMixin, RouteMixin) {
 
   //determines if the use can save the file
   get disable_saving() {
-    return typeof this.file?.database_id !== 'undefined' || this.saving;
+    return (
+      this.saving ||
+      (this.file?.database_id !== undefined && !this.file.hasUnsavedChanges)
+    );
+  }
+
+  file_data(file: EvaluationFile | ProfileFile) {
+    if (file.hasOwnProperty('evaluation')) {
+      return _.get(file, 'evaluation.data');
+    }
+    return _.get(file, 'profile.data');
+  }
+
+  update_database_file(file: EvaluationFile | ProfileFile) {
+    const evaluation = EvaluationModule.evaluationForFile(file) as
+      | IEvaluation
+      | undefined;
+    if (!evaluation) {
+      SnackbarModule.failure('Could not find the database entry for this file.');
+      return;
+    }
+
+    this.saving = true;
+    EvaluationModule.updateEvaluation({
+      ...evaluation,
+      data: this.file_data(file)
+    })
+      .then(() => {
+        SnackbarModule.notify('File saved successfully');
+        InspecDataModule.markFileSaved(file.uniqueId);
+      })
+      .catch((error) => {
+        SnackbarModule.HTTPFailure(error);
+      })
+      .finally(() => {
+        this.saving = false;
+      });
   }
 
   save_to_database(file: EvaluationFile | ProfileFile) {
@@ -116,26 +170,18 @@ export default class SidebarFileList extends mixins(ServerMixin, RouteMixin) {
       }
     }
     // Add evaluation data to the form
-    if (file.hasOwnProperty('evaluation')) {
-      formData.append(
-        'data',
-        new Blob([JSON.stringify(_.get(file, 'evaluation.data'))], {
-          type: 'text/plain'
-        })
-      );
-    } else {
-      formData.append(
-        'data',
-        new Blob([JSON.stringify(_.get(file, 'profile.data'))], {
-          type: 'text/plain'
-        })
-      );
-    }
+    formData.append(
+      'data',
+      new Blob([JSON.stringify(this.file_data(file))], {
+        type: 'text/plain'
+      })
+    );
     axios
       .post<IEvaluation>('/evaluations', formData)
       .then((response) => {
         SnackbarModule.notify('File saved successfully');
         file.database_id = parseInt(response.data.id);
+        InspecDataModule.markFileSaved(file.uniqueId);
         EvaluationModule.loadEvaluation(response.data.id);
         const loadedDatabaseIds = InspecDataModule.loadedDatabaseIds.join(',');
         this.navigateWithNoErrors(
